TemporaryFolder in .xtext.testing, deprecate the one in xbase.testing
